+++ src/potions/variableMeatPotion.ts
